mms_extractor/requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=2.0.0
sentence-transformers>=2.2.0
kiwipiepy>=0.15.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
openai>=1.0.0
anthropic>=0.25.0
httpx>=0.25.0
rapidfuzz>=3.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
scipy>=1.10.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
faiss-cpu>=1.7.4
psutil>=5.9.0
flask>=2.3.0
flask-cors>=4.0.0
